Feature/organization

+++ src/user/command/application/delete/delete.handler.ts
@@ -1,9 +1,9 @@
 import { Inject, Injectable } from '@nestjs/common';
-import { OAuthProviderFactory } from 'src/shared/core/infrastructure/oauth/oauth-provider.factory';
-import { OAuthProviderType } from 'src/auth/domain/value-object/oauth-provider.enum';
 import { USER_COMMAND_REPOSITORY, UserCommandRepository } from '../../domain/user.command.repository';
-import { CommandHandler, ICommandHandler } from '@nestjs/cqrs';
+import { CommandHandler, EventBus, ICommandHandler } from '@nestjs/cqrs';
 import { DeleteMyInfoCommand } from './delete.command';
+import { CustomException } from 'src/shared/exception/custom-exception';
+import { CustomExceptionCode } from 'src/shared/exception/custom-exception-code';
 
 @Injectable()
 @CommandHandler(DeleteMyInfoCommand)
@@ -11,15 +11,21 @@ export class DeleteMyInfoHandler implements ICommandHandler<DeleteMyInfoCommand>
   constructor(
     @Inject(USER_COMMAND_REPOSITORY)
     private readonly userCommandRepository: UserCommandRepository,
-    private readonly oAuthProviderFactory: OAuthProviderFactory,
+    private readonly eventBus: EventBus,
   ) {}
 
   async execute(command: DeleteMyInfoCommand): Promise<void> {
     const { userId } = command;
-    const oAuthProvider = this.oAuthProviderFactory.getProvider(OAuthProviderType.KAKAO);
-    console.log(1);
-    await oAuthProvider.unlinkAccount(userId);
+
+    const user = await this.userCommandRepository.findById(userId);
+    if (!user) {
+      throw new CustomException(CustomExceptionCode.USER_NOT_FOUND);
+    }
+
+    user.delete();
 
     await this.userCommandRepository.deleteById(userId);
+
+    await this.eventBus.publishAll(user.pullDomainEvents());
   }
 }
